+++ Jellyfin.Plugin.AzureIllusion/ScheduledTasks/SubtitleTaskExecutionGate.cs
@@ -1,0 +1,6 @@
+namespace Jellyfin.Plugin.AzureIllusion.ScheduledTasks;
+
+internal static class SubtitleTaskExecutionGate
+{
+    internal static SemaphoreSlim Instance { get; } = new(1, 1);
+}